refactor(persistence): split DB into submodules; move background sync to services; fix(cors): allow x-api-key
