+++ .gitignore
@@ -13,3 +13,4 @@ static/
 .codex
 test
 scripts
+samples
