+++ src/services/mod.rs
@@ -2,3 +2,4 @@ pub mod cookie_actor;
 pub mod key_actor;
 #[cfg(feature = "portable")]
 pub mod update;
+pub mod sync;
